Second audit: harden PNG input, de-duplicate the sync front end, v1.7.0
A full-project audit after 37 sessions, and the release of the two
features that had accumulated unreleased on main (PNG I/O, WMO-phasing
stations) together with what the audit found.

The one real defect was in the new PNG reader. The scanline length was
computed in `unsigned long`, which is 32 bits on MSVC and 64 on
macOS/Linux, while unfilter() indexes the same buffer in size_t. A file
declaring 100000x42950 grayscale wraps that product to 75,654 — so the
buffer is allocated at 75 KB and then written through for 4.3 GB, out of
a ~1 KB file, on Windows only. It is now computed in uint64_t and bounded
before anything is allocated; the bound also keeps zlib's own 32-bit
uLongf well defined. bmp_read got the same treatment: benign there today
because an oversized allocation throws first, but the bounds check is
what should be load-bearing, not the allocator.

The reader also skipped straight past every chunk CRC. That is the one
property making PNG worth calling the archival auto-save format, so bit
rot in a saved chart came back as quiet garbage. Critical chunks are now
verified before their contents are used; a damaged ancillary chunk is
still ignored, as the spec allows.

The audit's other theme was duplication. A normalized clone scan put 70%
of every duplicated block in the project in one place: live.cpp against
syncscan.cpp, which had been maintained as hand-written twins and cost
two same-day bugs when the WMO-phasing work landed. The tree already
contained the answer one layer down — FmDecoder is a streaming class and
batch fm_decode() is a loop over it — so the sharing was extended:
sync_ma_step (the moving average's ramp-up divisor), SyncRuns (one run
rule for both polarities, replacing four copies) and both fallback
searches now live in syncscan.h and are called by both paths. The GUI's
print renderer turned out to be the export renderer at scale 1, box-
averaging a 1x1 box being the identity. About 200 lines of parallel
algorithm removed, with all eight recordings decoding byte-identical to
a baseline taken first.

The export/print render moved to core/render.* because it was static
inside gui/main.cpp where nothing headless could reach it, and none of
it is FLTK. render-test now pins the 2280-line print ruler — printing a
partial reception once came out ~2x stretched on real paper because the
page height was scaled by lines received instead of by the original's
fixed drum.

Smaller: out-of-range ini integers fall back instead of wrapping into a
plausible-looking setting (5000000000 became 705032704, which then clamps
downstream and looks deliberate); syn_write verifies line width rather
than trusting callers; the dev-only --test-* indices are clamped; and
isobar-decode recognises .SYN as well as .syn.

Every new assertion was run against the old code first and had to fail
with the right message — a test that only passes with the fix in place
proves nothing. ctest 21 -> 22, warning-free, all fixtures unchanged.

Two features unreleased since v1.6.0 make this a minor bump.

+++ core/bmpfile.cpp
@@ -106,14 +106,23 @@ void bmp_read(const std::string &path, std::vector<uint8_t> &rgb,
     int top_down = h < 0;
     if (top_down) h = -h;
     int bpp = bits / 8;
-    unsigned long stride = ((unsigned long)w * bpp + 3) & ~3UL;
-    if (data_off + stride * (unsigned long)h > file.size())
+    /* 64-bit throughout. `unsigned long` is 32 bits on MSVC and 64 on
+     * macOS/Linux, so stride*h computed in it wraps on Windows for a large
+     * enough declared size - the bounds check would then pass and the row
+     * loop below, which indexes in size_t, would read past the file buffer.
+     * (The oversized rgb.assign() happens to throw first today, but the
+     * check is what is meant to be load-bearing here, not the allocator.
+     * core/pngfile.cpp documents the same trap at PNG_MAX_RAW, where it
+     * really did bite.) */
+    uint64_t stride = ((uint64_t)w * bpp + 3) & ~(uint64_t)3;
+    if ((uint64_t)data_off + stride * (uint64_t)h > (uint64_t)file.size())
         throw std::runtime_error("bmp_read: truncated file");
 
     rgb.assign((size_t)w * h * 3, 0);
     for (long y = 0; y < h; y++) {
         long sy = top_down ? y : h - 1 - y;
-        const uint8_t *src = &file[data_off + (size_t)sy * stride];
+        const uint8_t *src =
+            &file[(size_t)((uint64_t)data_off + (uint64_t)sy * stride)];
         uint8_t *dst = &rgb[(size_t)y * w * 3];
         for (long x = 0; x < w; x++) {
             dst[3 * x]     = src[bpp * x + 2];   /* R */
